Fix upload image API
