+++ backend/src/modules/file/controllers/post.file.controller.ts
@@ -1,35 +1,14 @@
-import {
-  Controller,
-  Post,
-  UseInterceptors,
-  UploadedFile,
-} from '@nestjs/common';
-import { diskStorage } from 'multer';
-import { extname } from 'path';
-import { FileInterceptor } from '@nestjs/platform-express';
+import { Controller, Post, Body } from '@nestjs/common';
 import { PostFileService } from '../services/post.file.service';
+import { FileEntity } from '../domain/file.entity';
+
 
 @Controller('file')
 export class PostFileController {
-  SERVER_URL: string = 'http://localhost:9000/';
   constructor(private postFileService: PostFileService) {}
 
   @Post('')
-  @UseInterceptors(
-    FileInterceptor('file', {
-      storage: diskStorage({
-        destination: './data',
-        filename: (req, file, cb) => {
-          const randomName = Array(32)
-            .fill(null)
-            .map(() => Math.round(Math.random() * 16).toString(16))
-            .join('');
-          return cb(null, `${randomName}${extname(file.originalname)}`);
-        },
-      }),
-    }),
-  )
-  uploadAvatar(@UploadedFile() file) {
-    this.postFileService.create(`${this.SERVER_URL}${file.path}`);
+  async create(@Body() body: FileEntity) {
+    return await this.postFileService.create(body);
   }
 }
