+++ backend/src/modules/file/controllers/upload.file.controller.ts
@@ -1,0 +1,38 @@
+import {
+  Controller,
+  Post,
+  Body,
+  UseInterceptors,
+  UploadedFile,
+} from '@nestjs/common';
+import { diskStorage } from 'multer';
+import { extname } from 'path';
+import { FileInterceptor } from '@nestjs/platform-express';
+
+@Controller('file')
+export class UploadFileController {
+  SERVER_URL = 'http://localhost:9000/';
+
+  @Post('upload')
+  @UseInterceptors(
+    FileInterceptor('file', {
+      storage: diskStorage({
+        destination: './data',
+        filename: (req, file, cb) => {
+          const randomName = Array(32)
+            .fill(null)
+            .map(() => Math.round(Math.random() * 16).toString(16))
+            .join('');
+          return cb(null, `${randomName}${extname(file.originalname)}`);
+        },
+      }),
+    }),
+  )
+  async uploadImage(@UploadedFile() file) {
+    const response = {
+      fileURL: `${this.SERVER_URL}${file.filename}`,
+      file: file,
+    };
+    return response;
+  }
+}
